LayoutTests/http/tests/security/canvas-remote-read-remote-video-hls-expected.txt

CONSOLE MESSAGE: line 1: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
Ensure that data cannot be retrieved from a canvas tainted by a video resource with remote subresources when CORS is not enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing data retrieval on an untainted canvas:
PASS canvas.getContext('2d').getImageData(0, 0, 100, 100) did not throw exception.
PASS canvas.toDataURL() did not throw exception.

Testing data retrieval on a canvas tainted by a pattern generated by a hls video resource containing remote contents:
PASS context.getImageData(0, 0, 100, 100) threw exception SecurityError: The operation is insecure..
PASS canvas.toDataURL() threw exception SecurityError: The operation is insecure..
PASS successfullyParsed is true

TEST COMPLETE
